Port the Windows parts to D2.

+++ dvm/dvm/Exceptions.d
@@ -12,7 +12,7 @@ class DvmException : Exception
 {
 	template Constructor ()
 	{
-		this (string message, string file, long line)
+		this (string message, string file, size_t line)
 		{
 			super(message, file, line);
 		}
